chore(feedback): feedback implementation - add prompt on delete user - revisit the reducer - fix wrong error message in the api route - remove commented code - adjust default export to match file names - create a helper file to handle tokens and test datas - rename test files to match original filenames - fix modal footer buttons to take full width - add title to the admin bar
