+++ .codeclimate.yml
@@ -14,3 +14,4 @@ ratings:
   - "**.js"
 exclude_paths:
 - "node_modules/"
+- "client/assets/"
