+++ .coveralls.yml
@@ -1,0 +1,1 @@
+repo_token: m6wyhr3bfJNKeSVygqUGOjr89QTw7H04D
