+++ .jshintignore
@@ -1,0 +1,2 @@
+client/assets/
+node_modules
