+++ .travis.yml
@@ -10,3 +10,6 @@ script:
 addons:
     code_climate:
         repo_token: cd4494dc1d820cda82620359978bec8d1a4a82ba8fa249bb0448e879e88ad0e9
+env:
+  CODECLIMATE_REPO_TOKEN: cd4494dc1d820cda82620359978bec8d1a4a82ba8fa249bb0448e879e88ad0e9
+after_success: npm run coverage
