+++ README.md
@@ -1,4 +1,5 @@
 # DOKUMENT &trade;
 [![Build Status](https://travis-ci.org/andela-dbamidele/checkpoint2.svg?branch=master)](https://travis-ci.org/andela-dbamidele/checkpoint2)
+[![Coverage Status](https://coveralls.io/repos/github/andela-dbamidele/checkpoint2/badge.svg?branch=master)](https://coveralls.io/github/andela-dbamidele/checkpoint2?branch=master)
 
 More info comming soon...!
