--- circle.yml
@@ -1,6 +1,0 @@
-machine:
-  node
-
-test:
-  pre:
-    - npm install -g sequelize-cli
