+++ client/__tests__/components/body/document/TinyMceComponent.spec.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import { mount } from 'enzyme';
+import TinyMceComponent from
+  '../../../../components/body/document/TinyMceComponent';
+
+describe('Tinymce Editor Component', () => {
+  const handleEditorChange = jest.fn(() => null);
+  const wrapper = mount(
+    <TinyMceComponent
+      id="hello"
+      handleEditorChange={handleEditorChange}
+      content="Hello world"
+    />
+  );
+  it('renders', () => {
+    expect(wrapper.find('#hello'))
+    .toHaveLength(1);
+    wrapper.instance().componentDidMount();
+  });
+});
