+++ client/__tests__/components/body/landing/Editing.spec.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import Editing from '../../../../components/body/landing/Editing';
+
+describe('Landing column 1', () => {
+  const wrapper = shallow(
+    <Editing />
+  );
+  it('renders', () => {
+    expect(wrapper.find('#home1'))
+    .toHaveLength(1);
+  });
+});
