+++ client/assets/sass/dokuments.scss
@@ -428,6 +428,10 @@ nav {
   }
 }
 
+.modal-footer span {
+  min-width: 100px !important;
+}
+
 .button-collapse i {
   color: #000;
 }
@@ -560,6 +564,9 @@ input:-webkit-autofill {
 .mt-20 {
   margin-top: 20px;
 }
+.pt-10 {
+  padding-top: 10px;
+}
 #nav-mobile li {
   padding-left: 5px !important;
   padding-right: 5px !important;
@@ -575,6 +582,10 @@ input:-webkit-autofill {
 
 #errorImg {
   max-height: 300px !important;
+}
+
+button.swal2-confirm {
+  margin-right: 10px;
 }
 
 @media only screen and (max-width : 600px) {
