+++ client/components/body/Landing.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import Parallax from './landing/Parallax';
-import HomeColumn1 from './landing/Editing';
-import HomeColumn2 from './landing/Security';
-import HomeColumn3 from './landing/GetStarted';
+import Editing from './landing/Editing';
+import Security from './landing/Security';
+import GetStarted from './landing/GetStarted';
 
 const Landing = () => (
   <div>
     <Parallax />
-    <HomeColumn1 />
-    <HomeColumn2 />
-    <HomeColumn3 />
+    <Editing />
+    <Security />
+    <GetStarted />
   </div>
 );
