+++ server/__tests__/helpers/token.json
@@ -1,0 +1,4 @@
+{
+  "admin": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJib2x1d2F0aWZlbWkiLCJpYXQiOjE0OTgyODgyNDl9.PtsWTssgKdF5vtqJl6rNG4S9-4XxD7rBK3n3sgwEhUQ",
+  "regular": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MiwidXNlcm5hbWUiOiJiaW50aTQiLCJmdWxsbmFtZSI6IkVtbWFudWVsbGEiLCJyb2xlSWQiOjIsImlhdCI6MTQ5OTk3MTQwOH0.7F0YTUWb4JJwhbWyNE-Mfaax4D_5124i1OLxFg-8OYc"
+}
