+++ server/__tests__/users.routes.spec.js
@@ -3,6 +3,7 @@ import chaiHttp from 'chai-http';
 import supertest from 'supertest';
 import isEmpty from 'lodash/isEmpty';
 import app from '../server';
+import token from './helpers/token.json';
 
 const User = require('../models').User;
 // const Document = require('../models').Document;
@@ -10,10 +11,6 @@ const Role = require('../models').Role;
 
 const request = supertest.agent(app);
 chai.use(chaiHttp);
-
-const Auth = 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6' +
-              'MSwidXNlcm5hbWUiOiJib2x1d2F0aWZlbWkiLCJpYXQiO' +
-              'jE0OTgyODgyNDl9.PtsWTssgKdF5vtqJl6rNG4S9-4XxD7rBK3n3sgwEhUQ';
 
 process.env.NODE_ENV = 'test';
 
@@ -73,7 +70,7 @@ describe('API Routes', () => {
         request
         .get('/api/users')
         .set('Accept', 'application/json')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.status).to.equal(200);
@@ -86,7 +83,7 @@ describe('API Routes', () => {
         request
         .get('/api/users')
         .set('Accept', 'application/json')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.body.users).to.be.an('array');
@@ -124,7 +121,7 @@ describe('API Routes', () => {
         };
         request
         .post('/api/users')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .send(user)
         .end((err, res) => {
           if (!err) {
@@ -151,7 +148,7 @@ describe('API Routes', () => {
 
         request
         .post('/api/users/')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .send(user)
         .end((err, res) => {
           if (!err) {
@@ -174,7 +171,7 @@ describe('API Routes', () => {
         };
         request
         .post('/api/users')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .send(user)
         .end((err, res) => {
           if (!err) {
@@ -206,7 +203,7 @@ describe('API Routes', () => {
       it('should get a user by id', (done) => {
         request
         .get('/api/users/1/')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.status).to.equal(200);
@@ -222,7 +219,7 @@ describe('API Routes', () => {
       it('should return status 404 if user is not found', (done) => {
         request
         .get('/api/users/10')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.status).to.equal(404);
@@ -235,7 +232,7 @@ describe('API Routes', () => {
       it('should return error if the userid is not a number', (done) => {
         request
         .get('/api/users/ade')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.status).to.equal(400);
@@ -267,7 +264,7 @@ describe('API Routes', () => {
       it('return error if the userid is not a number', (done) => {
         request
         .put('/api/users/ade')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.status).to.equal(400);
@@ -280,7 +277,7 @@ describe('API Routes', () => {
       it('return error if the user is not found', (done) => {
         request
         .put('/api/users/10')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.status).to.equal(400);
@@ -293,7 +290,7 @@ describe('API Routes', () => {
       it('edits users', (done) => {
         request
         .put('/api/users/1/')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .send({
           fullname: 'Ibukunoluwa',
         })
@@ -327,7 +324,7 @@ describe('API Routes', () => {
       it('return error if the userid is not a number', (done) => {
         request
         .delete('/api/users/ade')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.status).to.equal(400);
@@ -340,7 +337,7 @@ describe('API Routes', () => {
       it('deletes user by id', (done) => {
         request
         .delete('/api/users/1')
-        .set('Authorization', Auth)
+        .set('Authorization', token.admin)
         .end((err, res) => {
           if (!err) {
             expect(res.status).to.equal(200);
